Fix freezing model error
This was not repeatable on LambdaLabs previously, only on replicate;
however today I was able to see it freezing on LambdaLabs as well while
processing our example 5m audio file. Updating to the current version of
stable-ts fixed the freezing issue, though I don't know why 🤷.

Using the cpu pre-loading resulted in an encoding error for Japanese,
so we remove that, too. Presumably that's a bug in stable-ts or Whisper.